+++ src/main/java/kitchenpos/common/infra/PurgomalumClient.java
@@ -1,0 +1,6 @@
+package kitchenpos.common.infra;
+
+public interface PurgomalumClient {
+
+    boolean containsProfanity(String text);
+}
