--- src/main/java/kitchenpos/domain/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.domain;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
