--- src/main/java/kitchenpos/infra/PurgomalumClient.java
@@ -1,5 +1,0 @@
-package kitchenpos.infra;
-
-public interface PurgomalumClient {
-    boolean containsProfanity(String text);
-}
