+++ src/main/java/kitchenpos/order/common/OrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.common;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
