+++ src/main/java/kitchenpos/order/delivery/domain/DeliveryOrderRepository.java
@@ -1,0 +1,15 @@
+package kitchenpos.order.delivery.domain;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface DeliveryOrderRepository {
+
+    DeliveryOrder save(DeliveryOrder order);
+
+    Optional<DeliveryOrder> findById(UUID id);
+
+    List<DeliveryOrder> findAll();
+}
+
