+++ src/main/java/kitchenpos/order/delivery/domain/DeliveryOrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.delivery.domain;
+
+public enum DeliveryOrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
