+++ src/main/java/kitchenpos/order/delivery/infra/KitchenridersClient.java
@@ -1,0 +1,9 @@
+package kitchenpos.order.delivery.infra;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
