+++ src/main/java/kitchenpos/order/delivery/ui/DeliveryOrderRestController.java
@@ -1,0 +1,63 @@
+package kitchenpos.order.delivery.ui;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+import kitchenpos.order.delivery.application.DeliveryOrderService;
+import kitchenpos.order.delivery.domain.DeliveryOrder;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/orders/delivery")
+@RestController
+public class DeliveryOrderRestController {
+
+    private final DeliveryOrderService orderService;
+
+    public DeliveryOrderRestController(final DeliveryOrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<DeliveryOrder> create(@RequestBody final DeliveryOrder request) {
+        final DeliveryOrder response = orderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/delivery/" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<DeliveryOrder> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<DeliveryOrder> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/start-delivery")
+    public ResponseEntity<DeliveryOrder> startDelivery(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.startDelivery(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete-delivery")
+    public ResponseEntity<DeliveryOrder> completeDelivery(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.completeDelivery(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<DeliveryOrder> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<DeliveryOrder>> findAll() {
+        return ResponseEntity.ok(orderService.findAll());
+    }
+}
