+++ src/main/java/kitchenpos/order/eatin/domain/EatInOrderRepository.java
@@ -1,0 +1,17 @@
+package kitchenpos.order.eatin.domain;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface EatInOrderRepository {
+
+    EatInOrder save(EatInOrder order);
+
+    Optional<EatInOrder> findById(UUID id);
+
+    List<EatInOrder> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, EatInOrderStatus status);
+}
+
