+++ src/main/java/kitchenpos/order/eatin/domain/EatInOrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.eatin.domain;
+
+public enum EatInOrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
