+++ src/main/java/kitchenpos/order/eatin/domain/JpaEatInOrderRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.order.eatin.domain;
+
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaEatInOrderRepository extends EatInOrderRepository, JpaRepository<EatInOrder, UUID> {
+}
