+++ src/main/java/kitchenpos/order/eatin/ui/EatInOrderRestController.java
@@ -1,0 +1,53 @@
+package kitchenpos.order.eatin.ui;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+import kitchenpos.order.eatin.application.EatInOrderService;
+import kitchenpos.order.eatin.domain.EatInOrder;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/orders/eat-in")
+@RestController
+public class EatInOrderRestController {
+
+    private final EatInOrderService orderService;
+
+    public EatInOrderRestController(final EatInOrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<EatInOrder> create(@RequestBody final EatInOrder request) {
+        final EatInOrder response = orderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/eat-in/" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<EatInOrder> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<EatInOrder> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<EatInOrder> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<EatInOrder>> findAll() {
+        return ResponseEntity.ok(orderService.findAll());
+    }
+}
