+++ src/main/java/kitchenpos/order/takeout/domain/JpaTakeOutOrderRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.order.takeout.domain;
+
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaTakeOutOrderRepository extends TakeOutOrderRepository,
+    JpaRepository<TakeOutOrder, UUID> {
+
+}
