+++ src/main/java/kitchenpos/order/takeout/domain/TakeOutOrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.takeout.domain;
+
+public enum TakeOutOrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
