+++ src/main/java/kitchenpos/order/takeout/ui/TakeOutOrderRestController.java
@@ -1,0 +1,53 @@
+package kitchenpos.order.takeout.ui;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+import kitchenpos.order.takeout.application.TakeOutOrderService;
+import kitchenpos.order.takeout.domain.TakeOutOrder;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/orders/take-out")
+@RestController
+public class TakeOutOrderRestController {
+
+    private final TakeOutOrderService orderService;
+
+    public TakeOutOrderRestController(final TakeOutOrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<TakeOutOrder> create(@RequestBody final TakeOutOrder request) {
+        final TakeOutOrder response = orderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/take-out" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<TakeOutOrder> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<TakeOutOrder> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<TakeOutOrder> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(orderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<TakeOutOrder>> findAll() {
+        return ResponseEntity.ok(orderService.findAll());
+    }
+}
