+++ src/test/java/kitchenpos/application/FakeKitchenridersClient.java
@@ -1,17 +1,17 @@
 package kitchenpos.application;
 
-import kitchenpos.infra.KitchenridersClient;
-
 import java.math.BigDecimal;
 import java.util.UUID;
 
 public class FakeKitchenridersClient implements KitchenridersClient {
+
     private UUID orderId;
     private BigDecimal amount;
     private String deliveryAddress;
 
     @Override
-    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
+    public void requestDelivery(final UUID orderId, final BigDecimal amount,
+        final String deliveryAddress) {
         this.orderId = orderId;
         this.amount = amount;
         this.deliveryAddress = deliveryAddress;
