+++ src/test/java/kitchenpos/application/FakePurgomalumClient.java
@@ -1,11 +1,11 @@
 package kitchenpos.application;
 
-import kitchenpos.infra.PurgomalumClient;
-
 import java.util.Arrays;
 import java.util.List;
+import kitchenpos.common.infra.PurgomalumClient;
 
 public class FakePurgomalumClient implements PurgomalumClient {
+
     private static final List<String> profanities;
 
     static {
