+++ src/test/java/kitchenpos/application/InMemoryOrderRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.application;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderRepository;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.eatin.domain.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
